refactor: separate game scenes.
