--- src/game.py
@@ -1,6 +1,0 @@
-class Game:
-    def __init__(self):
-        pass
-
-    def run(self):
-        pass
