+++ src/scenes/game.py
@@ -1,0 +1,9 @@
+from .scene import Scene
+
+class Game(Scene):
+    def __init__(self):
+        pass
+
+    def run(self, screen):
+        pass
+
