+++ src/scenes/scene.py
@@ -1,0 +1,3 @@
+class Scene:
+    def run(self, screen):
+        raise NotImplementedError()
